requirements.txt
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
google-generativeai>=0.5.0
pydantic>=2.0.0
pyyaml>=6.0.0
